refactor(github): extract issue-ops (view/create/edit/comment/list/detect-linked-pr) into @dev-loops/core
Moves the gh-command logic out of the thin scripts/github/*.mjs CLI
wrappers into packages/core/src/github/issue-ops.mjs, mirroring the
existing move-queue-item/list-queue-items core+thin-wrapper split. Pure
extraction (behavior-preserving) that lets the upcoming tracker seam's
GitHub provider adapter call one implementation instead of duplicating
gh-command construction.
